+++ demos/manual/src/main/java/is/codion/manual/design/DesignAndStyle.java
@@ -70,7 +70,8 @@ public final class DesignAndStyle {
 										.build();
 
 		TemporalField<LocalDate> field =
-						TemporalField.builder(LocalDate.class)
+						TemporalField.builder()
+										.temporalClass(LocalDate.class)
 										.dateTimePattern("dd.MM.yyyy")
 										.columns(12)
 										.border(createTitledBorder("Date"))